+++ modules/dnn/src/vkcom/src/op_base.cpp
@@ -16,8 +16,8 @@ namespace cv { namespace dnn { namespace vkcom {
 
 OpBase::OpBase()
 {
-    ctx_ = getContext();
-    device_ = ctx_->device;
+    createContext();
+    device_ = kDevice;
     pipeline_ = VK_NULL_HANDLE;
     cmd_buffer_ = VK_NULL_HANDLE;
     descriptor_pool_ = VK_NULL_HANDLE;
@@ -139,7 +139,7 @@ void OpBase::createCommandBuffer()
 {
     VkCommandBufferAllocateInfo info = {};
     info.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_ALLOCATE_INFO;
-    info.commandPool = ctx_->cmd_pool;
+    info.commandPool = kCmdPool;
     info.level = VK_COMMAND_BUFFER_LEVEL_PRIMARY;
     info.commandBufferCount = 1;
     VK_CHECK_RESULT(vkAllocateCommandBuffers(device_, &info, &cmd_buffer_));
@@ -176,7 +176,7 @@ void OpBase::runCommandBuffer()
     fence_create_info_.flags = 0;
 
     VK_CHECK_RESULT(vkCreateFence(device_, &fence_create_info_, NULL, &fence));
-    VK_CHECK_RESULT(vkQueueSubmit(ctx_->queue, 1, &submit_info, fence));
+    VK_CHECK_RESULT(vkQueueSubmit(kQueue, 1, &submit_info, fence));
     VK_CHECK_RESULT(vkWaitForFences(device_, 1, &fence, VK_TRUE, 100000000000));
     vkDestroyFence(device_, fence, NULL);
 }
